chore: rename AccountsColumns composable
